refactor(coin): Command와 Query로 분리
